% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_internals.R
\name{course_internals}
\alias{course_internals}
\title{course_internals
We use page as an argument to lesson_intro because we can
have a second argument that defines a course}
\usage{
course_internals(page)
}
\description{
course_internals
We use page as an argument to lesson_intro because we can
have a second argument that defines a course
}
