% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_internals.R
\name{course_internals_basic}
\alias{course_internals_basic}
\title{course_internals_basic
We use page as an argument to lesson_intro because we can
have a second argument that defines a course}
\usage{
course_internals_basic(page)
}
\description{
course_internals_basic
We use page as an argument to lesson_intro because we can
have a second argument that defines a course
}
