% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_purrr.R
\name{make_page}
\alias{make_page}
\title{make_page}
\usage{
make_page(
  header = div("Nothing Here"),
  lesson_html = div(),
  code = "print(\\"Hello, world!\\")",
  display_editor = FALSE
)
}
\description{
make_page
}
