% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vim_tutor.R
\name{ndexModalDialog}
\alias{ndexModalDialog}
\title{ndexModalDialog}
\usage{
ndexModalDialog(
  ...,
  title = NULL,
  footer = modalButton("Dismiss"),
  size = c("m", "s", "l", "xl"),
  easyClose = FALSE,
  fade = TRUE
)
}
\description{
ndexModalDialog
}
